Rename module app to handler-notify-developer
