+++ handler-notify-developer/proguard-rules.pro
