+++ handler-notify-developer/src/androidTest/AndroidManifest.xml
@@ -8,7 +8,7 @@
         <activity android:name=".androidTest.TestActivity" />
 
         <activity
-            android:name=".NotifyDeveloperDialogDisplayActivity"
+            android:name=".handler.notifydeveloper.NotifyDeveloperDialogDisplayActivity"
             android:theme="@android:style/Theme.NoDisplay"/>
     </application>
 </manifest>
