+++ handler-notify-developer/src/androidTest/java/pl/brightinventions/slf4android/handler/notifydeveloper/TestActivity.java
@@ -1,4 +1,4 @@
-package pl.brightinventions.slf4android.androidTest;
+package pl.brightinventions.slf4android.handler.notifydeveloper;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -7,6 +7,6 @@ public class TestActivity extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.test);
+        setContentView(pl.brightinventions.slf4android.handler.notifydeveloper.androidTest.R.layout.test);
     }
 }
