+++ handler-notify-developer/src/androidTest/res/layout/test.xml
