+++ handler-notify-developer/src/main/AndroidManifest.xml
