+++ handler-notify-developer/src/main/java/pl/brightinventions/slf4android/handler/notifydeveloper/ActivityStateListener.java
@@ -1,4 +1,4 @@
-package pl.brightinventions.slf4android;
+package pl.brightinventions.slf4android.handler.notifydeveloper;
 
 import android.annotation.TargetApi;
 import android.app.Activity;
