+++ handler-notify-developer/src/main/java/pl/brightinventions/slf4android/handler/notifydeveloper/AtLeastFilter.java
@@ -1,7 +1,9 @@
-package pl.brightinventions.slf4android;
+package pl.brightinventions.slf4android.handler.notifydeveloper;
 
 import java.util.logging.Filter;
 import java.util.logging.LogRecord;
+
+import pl.brightinventions.slf4android.LogLevel;
 
 class AtLeastFilter implements Filter {
     private final LogLevel level;
