+++ handler-notify-developer/src/main/res/values/strings.xml
