+++ settings.gradle
@@ -2,7 +2,7 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.7.0"
 }
 
-include ':app'
 include ':core'
 include ':testapp'
 include ':handler-file-log'
+include ':handler-notify-developer'
